feat: gpt 5.5 went absolute HAM and fixed everything
note: .env file is now required, DO NOT push your application.properties with the credentials!

+++ backend/src/main/kotlin/org/hospitalmanagement/api/persons/DoctorController.kt
@@ -73,7 +73,7 @@ class DoctorController(private val doctorService: DoctorService) {
         @RequestParam(required = false) street: String?,
         @RequestParam(required = false) streetNo: Int?,
         @RequestParam(required = false) type: DoctorsType?,
-        @RequestParam(required = false) departmentId: Int?,
+        @RequestParam(required = false) departmentId: Long?,
         @RequestParam(required = false) workPhone: String?,
     ): Page<Doctor> {
         return doctorService.searchDoctors(
